limpeza do projeto
