--- TechChallenge.WebApp/ClientApp/src/app/home/home.component.html
@@ -1,106 +1,0 @@
-<!-- 
-<div style="display: flex; justify-content: end;">
-<div class="search">
-  <input
-  style="border-radius: 25px;"
-    type="text"
-    [(ngModel)]="term"
-    class="searchTerm"
-    placeholder="search"
-  />
-</div>
-</div>
-
-<div style="width: 15%;" *ngFor="let imagem of imagens">
-
-  <mat-card style="margin-top: 20px;" class="example-card">
-
-    <mat-card-header style="    display: flex;
-    justify-content: flex-end; margin: 0px;">
-      <mat-card-title style="display: flex;
-      align-items: center;">
-        <img class="fotoPerfil" src="{{imagem.urlPerfil}}" alt="Girl in a jacket">
-        <span>{{imagem.usuario}}</span>
-      </mat-card-title>
-      
-      
-    
-    </mat-card-header>
-
-    <mat-card-content>
-      <img style="height: 100%;
-      width: 100%;" src="{{imagem.foto.url}}" alt="Girl in a jacket">
-    </mat-card-content>
-    
-    <span>{{imagem.nome}}</span>
-    
-    <div style="display: flex;
-    justify-content: flex-end;">
-      <span style="font-size: 10px;">{{imagem.dataEnvio | date: 'dd/MM/yyyy'}}</span>
-    </div>
-
-    <mat-card-actions style="display: flex;">
-      <button mat-button>LIKE</button>
-      <button *ngIf="imagem.foto.publica" mat-button style="background-color: green;">
-        <span style="color: white;">SHARE</span>
-      </button>
-      <button *ngIf="!imagem.foto.publica" mat-button style="background-color: blue;">
-        <span style="color: white;">SHARE</span>
-      </button>
-    </mat-card-actions>
-    
-  </mat-card>
-</div>
-
-
-
-
-<input type="file" class="file-input" (change)="adicionar($event)" #fileUpload>
-
-<div style="display: flex;
-justify-content: end;" class="file-upload">
-
-  Selecione uma imagem
-
-  <button mat-mini-fab color="primary" class="upload-btn" (click)="fileUpload.click()">
-    <mat-icon>attach_file</mat-icon>
-  </button>
-</div> -->
-
-
-
-<div *ngIf="imagens.length == 0" class="page" style="background-color: #F5CEE6;">
-  <form style="display: flex; justify-content: center; align-items: center; width: 15%;height: 16%;" method="POST"
-    class="formLogin">
-    <mat-icon (click)="openDialog()"
-      style="transform: scale(6); position: absolute; top: 50%; left: 50%;">add_circle_outline</mat-icon>
-  </form>
-</div>
-
-
-
-
-<form style="background-color: white; display: flex; flex-wrap: wrap; flex-direction: column; padding: 2%;" *ngIf="imagens.length > 0">
-
-    <div name="botao" style="display: flex; justify-content: end;">
-      <button mat-raised-button color="primary" class="w-116" type="button" (click)="openDialog()">
-        Adicionar
-      </button>
-    </div>
-
-
-    <!-- <div style="width: 20%; height: 20%; display: flex; flex-wrap: wrap; flex-direction: column;" *ngFor="let imagem of imagens">
-      <app-publicacao [descricao]="imagem.nome" [url]="imagem.urlPerfil"></app-publicacao>
-    </div> --> 
-
-    <div style="width: 100%; display: flex; flex-wrap: wrap;">
-    <div style="width: 20%; border: solid; margin: 2%;" *ngFor="let imagem of imagens">
-      <app-publicacao style="overflow: hidden !important; display: flex; flex-direction: column; height: 100% !important;" [id]="imagem.id" [descricao]="imagem.nome" [url]="imagem.urlPerfil" (exluir)="excluirPublicacao($event)"></app-publicacao>
-    </div>
-  </div>
-</form>
-<!-- <div *ngIf="imagens.length > 0">
-<div *ngFor="let imagem of imagens">
-  {{imagem.url}}
-</div>
-</div> -->
