+++ TechChallenge.WebApp/ClientApp/src/app/models/Foto.ts
@@ -1,8 +1,14 @@
-// export class Foto{
-//     id:number;
-//     publica: boolean
-//     url: string;
-//     extensao:string;
-//     dataEnvio: Date;
-//     usuarioId: string;
-// }
+export class Foto {
+
+    constructor(url:string = null, publica:boolean = null, dataEnvio:Date = null)
+{
+    this.url = url;
+    this.publica = publica;
+    this.dataEnvio = dataEnvio;
+}
+    url:string;
+    publica:boolean;
+    dataEnvio:Date;
+    id:string;
+    extensao:string;
+}
