+++ TechChallenge.WebApp/ClientApp/src/app/models/Publicacao.ts
@@ -1,3 +1,5 @@
+import { Foto } from "./Foto";
+
 export class Publicacao {
     constructor(nome:string, usuario:string, urlPerfil:string, foto:Foto)
     {
@@ -13,17 +15,3 @@ export class Publicacao {
     id:number
 }
 
-export class Foto {
-
-    constructor(url:string = null, publica:boolean = null, dataEnvio:Date = null)
-{
-    this.url = url;
-    this.publica = publica;
-    this.dataEnvio = dataEnvio;
-}
-    url:string;
-    publica:boolean;
-    dataEnvio:Date;
-    id:string;
-    extensao:string;
-}
