+++ TechChallenge.WebApp/ClientApp/src/app/pages/Core/home/home.component.css
