+++ TechChallenge.WebApp/ClientApp/src/app/pages/Core/home/publicacao/nova-publicacao/nova-publicacao/nova-publicacao.component.html
@@ -1,20 +1,3 @@
-<!-- <form>
-    <div>
-        <input type="file" (change)="selectFile($event)" class="form-control">
-        <img *ngIf="localUrl" [src]="localUrl" class="imgPlaceholder">
-    </div>
-
-    <div style="display: flex; justify-content: center; margin-top: 5%;">
-        <input placeholder="descricao">
-    </div>
-
-    <div style="display: flex; justify-content: end;">
-        <input [disabled]="!localUrl" type="submit" (click)="uploadFile()" class="btn btn-success" value="Enviar">
-    </div>
-
-</form> -->
-
-
 <form [formGroup]="formGroup" method="POST" class="formLogin">
     <mat-horizontal-stepper #stepper>
         <mat-step>
@@ -39,11 +22,6 @@
             align-items: center;
             display: flex;" for='selecao-arquivo' (click)="fileUpload.click()
             ">selecione uma imagem</label>
-                    <!-- <input id='selecao-arquivo' style="display: none;" type="file" (change)="selectFile($event)" class="form-control"> -->
-
-
-                    <!-- <input type="file" (change)="selectFile($event)" class="form-control">
-            <input type="submit" (click)="uploadFile()" class="btn btn-success" value="Enviar"> -->
                 </div>
 
                 <div *ngIf="Foto && Foto.url">
@@ -71,11 +49,6 @@
             align-items: center;
             display: flex;" for='selecao-arquivo' (click)="fileUpload.click()
             ">selecione uma imagem</label>
-                    <!-- <input id='selecao-arquivo' style="display: none;" type="file" (change)="selectFile($event)" class="form-control"> -->
-
-
-                    <!-- <input type="file" (change)="selectFile($event)" class="form-control">
-            <input type="submit" (click)="uploadFile()" class="btn btn-success" value="Enviar"> -->
                 </div>
 
                 <div *ngIf="Foto && Foto.url">
@@ -102,13 +75,6 @@
                 <input [disabled]="!Foto.url" type="submit" (click)="uploadFile()" class="btn btn-success"
                     value="Publicar">
             </div>
-
-            <!-- TODO -->
-            <!-- no momento de cadastro do usuario vamos subir uma imagem temporariamente caso a foto nao seja vinculada a nenhum usuario em 1 semana vamos apagar essa foto 
-            basta criar uma propriedade privada na classe da foto que sera o id da pessoa e uma propriedade que dira se foi vinculada ou nao 
-
-            vincualada {get => return idPessoa.hasValue}
-            criar um servico para limpar as imagens que nao foram publicadas que vamos subir temporariamente  -->
         </mat-step>
     </mat-horizontal-stepper>
 </form>
