+++ TechChallenge.WebApp/ClientApp/src/app/pages/Core/home/publicacao/publicacao.component.html
@@ -1,11 +1,3 @@
-
-
-
-<!-- <div>
-    <img style=" height: 90%;     width: 10%; border-radius: 50%;" [src]="perfil" *ngIf="perfil" class="imgPlaceholder">
-    <span style="margin-left: 3%;">rafael</span>
-</div> -->
-
 <div *ngIf="url" style="width: 100%; margin-top: 3%;">
     <img style="width: 100%;     height: 200px;" [src]="url" *ngIf="url" class="imgPlaceholder">
 </div>
