+++ TechChallenge.WebApp/ClientApp/src/app/pages/Core/home/publicacao/publicacao.component.ts
@@ -1,6 +1,7 @@
 import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
 import { MatSnackBar } from '@angular/material';
-import { AccountService } from 'src/app/pages/authentication/shared/account.service';
+import { AccountService } from 'src/app/pages/Authentication/shared/account.service';
+
 
 @Component({
   selector: 'app-publicacao',
@@ -16,11 +17,6 @@ export class PublicacaoComponent implements OnInit {
   @Input() readonly:boolean;
 
   @Output() exluir: EventEmitter<number> = new EventEmitter<number>();
-  
-  nomeUsuario:string = 'Rafael';
-  dataPublicacao:Date;
-  tempo:string = 'ha duas horas';
-  perfil = "https://www.wikihow.com/images_en/thumb/d/db/Get-the-URL-for-Pictures-Step-2-Version-6.jpg/v4-460px-Get-the-URL-for-Pictures-Step-2-Version-6.jpg.webp";
 
   constructor( private _acountRepository:AccountService,  private _snackBar: MatSnackBar) { }
 
@@ -29,14 +25,11 @@ export class PublicacaoComponent implements OnInit {
 
   Excluir()
   {
-    debugger
     this._acountRepository.excluirPublicacao(this.id)
     .subscribe(
       data => 
       {
-        debugger
         this.exluir.emit(this.id);
-        // this._snackBar.open('Publicação excluida com sucesso!');
       },
       err =>
       {
