+++ TechChallenge.WebApp/ClientApp/src/app/pages/authentication/authentication/authentication.component.ts
@@ -13,20 +13,4 @@ export class AuthenticationComponent implements OnInit {
 
   ngOnInit() {
   }
-
-  // onSubmit(){
-  //   debugger
-  // }
-
-  // adicionar()
-  // {
-  //   debugger
-
-  //   //logar no sistema caso der certo colocar o token no cache 
-
-  //   //e redirecionar para pagina inicial 
-  //   this.localStorageService.setToken('teste');
-
-  //   this.router.navigate([''])
-  // }
 }
