+++ TechChallenge.WebApp/ClientApp/src/app/pages/authentication/create-acount/create-acount.component.ts
@@ -28,13 +28,11 @@ export class CreateAcountComponent implements OnInit {
   cadastrar()
   {
     this.accountService.Cadastrar(this.form.value).subscribe(result => {
-      debugger
       this._snackBar.open('Usuario Cadastrado com sucesso!');
 
       this.router.navigate([''])
     },
     err =>{
-      debugger
       this._snackBar.open(err.message);
     });
   }
