+++ TechChallenge.WebApp/ClientApp/src/app/pages/authentication/login/login.component.ts
@@ -2,10 +2,10 @@ import { Component, OnInit } from '@angular/core';
 import { FormControl, FormGroup} from '@angular/forms';
 import { Router } from '@angular/router';
 import { localStorageService } from 'src/app/shared/localStorageService';
-import { AccountService } from '../shared/account.service';
 import { MatDialog, MatSnackBar } from '@angular/material';
 import { Title } from '@angular/platform-browser';
-import { NovaPublicacaoComponent } from 'src/app/home/nova-publicacao/nova-publicacao/nova-publicacao.component';
+import { AccountService } from '../shared/account.service';
+
 @Component({
   selector: 'app-login',
   templateUrl: './login.component.html',
@@ -26,9 +26,7 @@ export class LoginComponent implements OnInit {
 
   adicionar()
   {
-    debugger
     this.accountService.logar(this.form.value).subscribe(result => {
-      debugger
       this.localStorageService.setToken(result.data);
       this.router.navigate([''])
     },
