+++ TechChallenge.WebApp/ClientApp/src/app/pages/authentication/shared/account.service.ts
@@ -1,7 +1,7 @@
 import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
-import { Foto, Publicacao } from 'src/app/home/model/Imagem';
+import { Foto } from 'src/app/models/Foto';
 import { ResponseModel } from 'src/app/models/Login';
 import { environment } from 'src/environments/environment';
 
@@ -21,12 +21,6 @@ export class AccountService {
     headers: new HttpHeaders({ 'Content-Type': 'multipart/form-data; boundary=something' })
   };
 
-  // headers =
-  // {
-  //   headers:new HttpHeaders(' Content-Disposition ', 'multipart/form-data')
-
-  // } 
-  
   constructor(private http: HttpClient) { }
 
   logar(formData)
@@ -42,11 +36,8 @@ export class AccountService {
   adicionarFoto(file:File) :  Observable<any> 
   {
     const form: FormData = new FormData();
-    console.log(file.name);
-    console.log(file);
+    
     form.append('formFile', file);
-
-   
 
     return this.http.post<any>(`${this.minimalUrl}/Upload`, form, this.httpOptionsImage);
   }
@@ -58,9 +49,7 @@ export class AccountService {
     console.log(file);
     form.append('formFile', file);
 
-      return this.http.post<any>(`${this.coreUrl}/Foto`, form);
-  
-      // return "https://www.wikihow.com/images_en/thumb/d/db/Get-the-URL-for-Pictures-Step-2-Version-6.jpg/v4-460px-Get-the-URL-for-Pictures-Step-2-Version-6.jpg.webp";
+    return this.http.post<any>(`${this.coreUrl}/Foto`, form);
   }
 
   publicar(body:any) : Observable<any>
